Listen for onsubmit to support IME in Safari
In Chrome, keydown fires before compositionend,
so if a user presses Enter to finish composition,
we can detect that we are composing (and
accordingly, avoid erroneously adding a newline).

However, in Safari, the keydown event is fired
after compositionend.
So, we cannot tell whether we are composing
when Enter is pressed.

As a workaround, we can listen for the onsubmit
event, which is only fired when the user presses
Enter while not composing.