Fix unresponsive buttons bug
Previously, when the settings icon received a
`blur` event, the settings menu would close.
This would happen _before_ the click event fired
for a settings menu item, which would cancel the
click event.
Thus, the buttons would never respond to clicks.
This commit fixes the bug by only closing the
menu if the mouse is not over the menu.